# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent & Device Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  leader: "leader",
  agent: "agent",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  active: "active",
  suspended: "suspended",
} as const;

export interface User {
  id: number;
  name: string;
  username: string;
  role: UserRole;
  status: UserStatus;
  leaderId?: number | null;
  leaderName?: string | null;
  createdAt: string;
}

export interface LoginResponse {
  user: User;
  message?: string;
}

export interface ChangePasswordRequest {
  currentPassword: string;
  newPassword: string;
}

export interface ResetPasswordRequest {
  newPassword: string;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  leader: "leader",
  agent: "agent",
} as const;

export interface CreateUserRequest {
  name: string;
  username: string;
  password: string;
  role: CreateUserRequestRole;
  leaderId?: number | null;
}

export interface UpdateUserRequest {
  name?: string;
  username?: string;
  leaderId?: number | null;
}

export type UpdateStatusRequestStatus =
  (typeof UpdateStatusRequestStatus)[keyof typeof UpdateStatusRequestStatus];

export const UpdateStatusRequestStatus = {
  active: "active",
  suspended: "suspended",
} as const;

export interface UpdateStatusRequest {
  status: UpdateStatusRequestStatus;
}

export interface MoveAgentRequest {
  newLeaderId: number;
}

export type DeviceStatus = (typeof DeviceStatus)[keyof typeof DeviceStatus];

export const DeviceStatus = {
  active: "active",
  pending: "pending",
  removed: "removed",
  blacklisted: "blacklisted",
} as const;

export interface Device {
  id: number;
  imei: string;
  agentId?: number | null;
  leaderId?: number | null;
  status: DeviceStatus;
  dateAdded: string;
  agentName?: string | null;
  leaderName?: string | null;
}

export interface AddDeviceRequest {
  imei: string;
}

export interface MoveDeviceRequest {
  toAgentId: number;
}

export interface AssignDeviceRequest {
  agentId: number;
}

export interface ImeiConflictResponse {
  error: string;
  imei: string;
  currentOwner?: User;
  device?: Device;
}

export interface DeviceSearchResult {
  device: Device;
  agent?: User;
  leader?: User;
}

export interface BlacklistEntry {
  id: number;
  imei: string;
  reason?: string | null;
  addedById: number;
  addedByName: string;
  createdAt: string;
}

export interface BlacklistRequest {
  imei: string;
  reason?: string | null;
}

export type DeviceRequestType =
  (typeof DeviceRequestType)[keyof typeof DeviceRequestType];

export const DeviceRequestType = {
  delete: "delete",
  transfer: "transfer",
} as const;

export type DeviceRequestStatus =
  (typeof DeviceRequestStatus)[keyof typeof DeviceRequestStatus];

export const DeviceRequestStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface DeviceRequest {
  id: number;
  type: DeviceRequestType;
  deviceId: number;
  deviceImei: string;
  fromAgentId?: number | null;
  fromAgentName?: string | null;
  toAgentId?: number | null;
  toAgentName?: string | null;
  status: DeviceRequestStatus;
  rejectionReason?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateRequestPayloadType =
  (typeof CreateRequestPayloadType)[keyof typeof CreateRequestPayloadType];

export const CreateRequestPayloadType = {
  delete: "delete",
  transfer: "transfer",
} as const;

export interface CreateRequestPayload {
  type: CreateRequestPayloadType;
  deviceId: number;
  toAgentId?: number | null;
}

export interface RejectRequestPayload {
  reason?: string | null;
}

export interface PerformanceEntry {
  rank: number;
  agentId: number;
  agentName: string;
  leaderId?: number | null;
  leaderName?: string | null;
  devicesCount: number;
  date: string;
}

export interface WeeklyPerformanceResponse {
  weekStart: string;
  weekEnd: string;
  isFinalized: boolean;
  rankings: PerformanceEntry[];
}

export interface ActivityLog {
  id: number;
  action: string;
  details?: string | null;
  userId?: number | null;
  userName?: string | null;
  createdAt: string;
}

export interface LogsResponse {
  logs: ActivityLog[];
  total: number;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  device_added: "device_added",
  request_pending: "request_pending",
  request_approved: "request_approved",
  request_rejected: "request_rejected",
  weekly_winner: "weekly_winner",
} as const;

export interface Notification {
  id: number;
  userId: number;
  title: string;
  message: string;
  type: NotificationType;
  read: boolean;
  createdAt: string;
}

export interface AdminDashboard {
  totalLeaders: number;
  totalAgents: number;
  totalDevices: number;
  todaySales: number;
  topAgentToday?: PerformanceEntry | null;
}

export interface LeaderDashboard {
  totalAgents: number;
  totalDevices: number;
  pendingRequests: number;
  recentActivity: ActivityLog[];
}

export interface AgentDashboard {
  myDevices: number;
  todaySales: number;
  weeklySales: number;
  rankToday?: number | null;
}

export type ListUsersParams = {
  role?: ListUsersRole;
  status?: ListUsersStatus;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  admin: "admin",
  leader: "leader",
  agent: "agent",
} as const;

export type ListUsersStatus =
  (typeof ListUsersStatus)[keyof typeof ListUsersStatus];

export const ListUsersStatus = {
  active: "active",
  suspended: "suspended",
} as const;

export type ListDevicesParams = {
  agentId?: number;
  leaderId?: number;
  status?: ListDevicesStatus;
};

export type ListDevicesStatus =
  (typeof ListDevicesStatus)[keyof typeof ListDevicesStatus];

export const ListDevicesStatus = {
  active: "active",
  pending: "pending",
  removed: "removed",
  blacklisted: "blacklisted",
} as const;

export type SearchDeviceParams = {
  imei: string;
};

export type ListRequestsParams = {
  status?: ListRequestsStatus;
  type?: ListRequestsType;
};

export type ListRequestsStatus =
  (typeof ListRequestsStatus)[keyof typeof ListRequestsStatus];

export const ListRequestsStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type ListRequestsType =
  (typeof ListRequestsType)[keyof typeof ListRequestsType];

export const ListRequestsType = {
  delete: "delete",
  transfer: "transfer",
} as const;

export type GetDailyPerformanceParams = {
  date?: string;
  leaderId?: number;
};

export type GetWeeklyPerformanceParams = {
  weekStart?: string;
  leaderId?: number;
};

export type ListLogsParams = {
  limit?: number;
  offset?: number;
  userId?: number;
};