# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Agent & Device Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List devices
      parameters:
        - name: agentId
          in: query
          schema:
            type: integer
        - name: leaderId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, pending, removed, blacklisted]
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: addDevice
      tags: [devices]
      summary: Add device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDeviceRequest"
      responses:
        "201":
          description: Device added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "409":
          description: IMEI already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImeiConflictResponse"
  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get device by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Delete device (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /devices/{id}/move:
    put:
      operationId: moveDevice
      tags: [devices]
      summary: Move device to another agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveDeviceRequest"
      responses:
        "200":
          description: Device moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}/assign:
    put:
      operationId: assignDevice
      tags: [devices]
      summary: Assign device to an agent (leader)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignDeviceRequest"
      responses:
        "200":
          description: Device assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/search:
    get:
      operationId: searchDevice
      tags: [devices]
      summary: Search device by IMEI
      parameters:
        - name: imei
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceSearchResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /blacklist:
    get:
      operationId: listBlacklist
      tags: [devices]
      summary: List blacklisted IMEIs
      responses:
        "200":
          description: Blacklisted IMEIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlacklistEntry"
    post:
      operationId: blacklistImei
      tags: [devices]
      summary: Add IMEI to blacklist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlacklistRequest"
      responses:
        "201":
          description: IMEI blacklisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlacklistEntry"
  /blacklist/{id}:
    delete:
      operationId: removeFromBlacklist
      tags: [devices]
      summary: Remove IMEI from blacklist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed from blacklist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user]
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, leader, agent]
        status:
          type: string
          enum: [active, suspended]
        leaderId:
          type: integer
          nullable: true
        leaderName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, username, role, status, createdAt]
...
    Device:
      type: object
      properties:
        id:
          type: integer
        imei:
          type: string
        agentId:
          type: integer
          nullable: true
        leaderId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [active, pending, removed, blacklisted]
        dateAdded:
          type: string
          format: date-time
        agentName:
          type: string
          nullable: true
        leaderName:
          type: string
          nullable: true
      required: [id, imei, status, dateAdded]
    AddDeviceRequest:
      type: object
      properties:
        imei:
          type: string
      required: [imei]
    MoveDeviceRequest:
      type: object
      properties:
        toAgentId:
          type: integer
      required: [toAgentId]
    AssignDeviceRequest:
      type: object
      properties:
        agentId:
          type: integer
      required: [agentId]
    ImeiConflictResponse:
      type: object
      properties:
        error:
          type: string
        imei:
          type: string
        currentOwner:
          $ref: "#/components/schemas/User"
        device:
          $ref: "#/components/schemas/Device"
      required: [error, imei]
    DeviceSearchResult:
      type: object
      properties:
        device:
          $ref: "#/components/schemas/Device"
        agent:
          $ref: "#/components/schemas/User"
        leader:
          $ref: "#/components/schemas/User"
      required: [device]
    BlacklistEntry:
      type: object
      properties:
        id:
          type: integer
        imei:
          type: string
        reason:
          type: string
          nullable: true
        addedById:
          type: integer
        addedByName:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, imei, addedById, addedByName, createdAt]
    BlacklistRequest:
      type: object
      properties:
        imei:
          type: string
        reason:
          type: string
          nullable: true
      required: [imei]
...